Set defaults using fetch
When `set`ing the three variables in the host application, eg `set
:systemd_use_sudo, true` in `deploy.rb`, it's not picked up by the tasks
in this gem. By using `fetch`, if it was set previously - use it. If
not, set the default.

Also - use "--no-pager" for the status command so that running status
properly exits once it's done, not keeping the SSH connection alive.